refactor: 인풋 검증 책임 ProductInputChecker타입으로 분리 구현
